LayoutTests/imported/w3c/web-platform-tests/svg/animations/animate-stop-currentcolor-expected.txt


FAIL Animating 'stop-color' from 'currentcolor' assert_equals: expected "rgb(0, 128, 0)" but got "rgba(0, 64, 0, 0.5)"
